pyproject.toml

[project]
name = "backend-template"
version = "0.0.1"
description = "Reusable backend template"
authors = [{ name = "Template Maintainer", email = "template@example.com" }]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.1,<2.0.0",
    "uvicorn>=0.25.0,<1.0.0",
    "fastapi>=0.109.1,<1.0.0",
    "gunicorn>=23.0.0,<24.0.0",
    "fastapi-event>=0.1.3,<0.2.0",
    "pythondi>=1.2.4,<2.0.0",
    "ujson>=5.9.0,<6.0.0",
    "click>=8.1.7,<9.0.0",
    "redis>=5.0.1,<6.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "pyjwt>=2.8.0,<3.0.0",
    "pytest-asyncio>=0.25.3,<1.0.0",
    "cryptography>=41.0.7,<42.0.0",
    "coverage>=7.4.0,<8.0.0",
    "sqlalchemy[asyncio]>=2.0.25,<3.0.0",
    "dependency-injector>=4.48.1,<5.0.0",
    "httpx>=0.26.0,<1.0.0",
    "dify-client>=0.1.10",
    "python-dotenv>=1.1.1",
    "dotenv>=0.9.9",
    "python-multipart>=0.0.20",
    "bcrypt>=4.3.0",
    "aioboto3>=15.1.0",
    "aiobotocore>=2.24.0",
    "boto3-stubs>=1.40.35",
    "types-aiobotocore>=2.24.2",
    "asyncpg>=0.30.0",
    "psycopg[binary]==3.2.10",
    "python-magic>=0.4.27",
    "markdown>=3.10",
    "pymdown-extensions>=10.17.1",
    "pillow",
    "temporalio>=1.21.1",
    "aiokafka>=0.13.0",
    "structlog>=24.2.0,<25.0.0",
    "asgi-correlation-id>=4.3.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0,<9.0.0",
    "types-redis>=4.6.0.20240106,<5.0.0",
]

[tool.coverage.run]
omit=["tests/*", "test_*.py", "migrations/*"]

[tool.isort]
profile="black"

[dependency-groups]
test = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.25.3",
    "pytest-pythonpath>=0.7.3",
    "testcontainers[kafka,redis]>=4.13.3",
]

# [build-system]
# requires = ["hatchling>=1.18"]
# build-backend = "hatchling.build"
